+++ src/adapter/response/AllArticlesResponse.ts
