+++ src/adapter/response/ArticleResponse.ts
