+++ src/application/article/ArticleApplicationService.ts
@@ -3,6 +3,7 @@ import { ArticleId } from '../../domain/models/article/valueObject/ArticleId'
 import { NotFoundError } from '../../http/errors/NotFoundError'
 import { HTTP_ERROR_MESSAGE } from '../../http/httpStatus'
 import { IArticleApplicationService } from './IArticleApplicationService'
+import { QueryParameters } from '../../adapter/request/QueryParameters'
 export class ArticleApplicationService implements IArticleApplicationService {
   constructor(private readonly articleRepository: IArticleRepository) {}
   public async get(id: string) {
@@ -14,13 +15,14 @@ export class ArticleApplicationService implements IArticleApplicationService {
     return article
   }
 
-  public async getAll() {
-    const articles = await this.articleRepository.findAll()
+  public async getAll(query: QueryParameters) {
+    const articles = await this.articleRepository.findAll(query)
     if (!articles) {
       throw new NotFoundError(HTTP_ERROR_MESSAGE.NotFound)
     }
     return articles
   }
+
   public async delete(id: string) {
     const articleId = new ArticleId(id)
     const article = await this.articleRepository.find(articleId)
