+++ src/application/article/IArticleApplicationService.ts
@@ -1,8 +1,9 @@
+import { QueryParameters } from '../../adapter/request/QueryParameters'
 import { Article } from '../../domain/models/article/entities/Article'
 import { Articles } from '../../domain/models/article/entities/Articles'
 
 export interface IArticleApplicationService {
   get(id: string): Promise<Article>
-  getAll(): Promise<Articles>
+  getAll(query: QueryParameters): Promise<Articles>
   delete(id: string): Promise<Article>
 }
