+++ src/bootstrap.ts
@@ -1,6 +1,6 @@
 import { ArticleRepository } from './repository/article/ArticleRepository'
 import { ArticleApplicationService } from './application/article/ArticleApplicationService'
-import { ArticleController } from './controller/ArticleController'
+import { ArticleController } from './adapter/controller/ArticleController'
 
 const repository = new ArticleRepository()
 const service = new ArticleApplicationService(repository)
